Made an own row sorter and made some fixes

+++ src/stats_lol/ClearData.java
@@ -3,6 +3,7 @@ package stats_lol;
 import java.io.File;
 import java.io.IOException;
 import java.net.URISyntaxException;
+import javax.swing.JOptionPane;
 import static stats_lol.Stats_lol.*;
 
 /**
@@ -12,30 +13,33 @@ import static stats_lol.Stats_lol.*;
 
 public class ClearData {
     
-    public void clearData() throws IOException, URISyntaxException{
+    public int clearData() throws IOException, URISyntaxException{
         
-        File plDt = new File("data/plData");
-        File tmDt = new File("data/tmData");
-        File chDt = new File("data/chData");
+        final int updateQ = JOptionPane.showConfirmDialog(null,"You will be ereasing all data, are you sure?","Clear data",JOptionPane.YES_NO_OPTION);
         
-        for(int i = 0; i < 40; i++){
-            File pl = new File(plDt, player[i]);
-            pl.delete();           
-        }       
-        
-        for(int i = 0; i < 8; i++){
-            File tm = new File(tmDt, team[i]);
-            tm.delete();
+        if(updateQ == 0){
             
+            File plDt = new File("data/plData");
+            File tmDt = new File("data/tmData");
+            File chDt = new File("data/chData");
+        
+            for(int i = 0; i < 40; i++){
+                File pl = new File(plDt, player[i]);
+                pl.delete();           
+            }       
+        
+            for(int i = 0; i < 8; i++){
+                File tm = new File(tmDt, team[i]);
+                tm.delete();
+            }
+        
+            for(int i = 0; i < 124; i++){
+                File ch = new File(chDt, champion[i]);
+                ch.delete();          
+            }
+        
+            Stats_lol.main(team);  
         }
-        
-        for(int i = 0; i < 124; i++){
-            File ch = new File(chDt, champion[i]);
-            ch.delete();          
-        }
-        
-        Stats_lol.main(team);
-        
-    }
-    
+        return updateQ;
+    }  
 }
